Updated i_tiny_spi to use buffers

+++ include/hal/i_tiny_async_i2c.h
@@ -31,7 +31,7 @@ typedef struct i_tiny_async_i2c_api_t {
     uint8_t address,
     bool prepare_for_restart,
     const uint8_t* buffer,
-    uint8_t buffer_size,
+    uint16_t buffer_size,
     tiny_async_i2c_callback_t callback,
     void* context);
 
@@ -48,7 +48,7 @@ typedef struct i_tiny_async_i2c_api_t {
     uint8_t address,
     bool prepare_for_restart,
     uint8_t* buffer,
-    uint8_t buffer_size,
+    uint16_t buffer_size,
     tiny_async_i2c_callback_t callback,
     void* context);
